add BigDecimal

+++ Module 1 Groovy/src/main/groovy/by/prohor/task1/calculate/expression/Addition.groovy
@@ -6,7 +6,7 @@ class Addition implements Expression {
     Expression right
 
     @Override
-    double eval(){
+    BigDecimal eval(){
         left.eval() + right.eval()
     }
 }
