+++ Module 1 Groovy/src/main/groovy/by/prohor/task1/calculate/expression/Expression.groovy
@@ -1,5 +1,5 @@
 package main.groovy.by.prohor.task1.calculate.expression
 
 interface Expression {
-    double eval()
+    BigDecimal eval()
 }
