+++ Module 1 Groovy/src/main/groovy/by/prohor/task1/calculate/expression/atom_value/ExpressionAtom.groovy
@@ -7,7 +7,7 @@ class ExpressionAtom extends Atom {
    Expression expression
 
     @Override
-    double eval() {
+    BigDecimal eval() {
         expression.eval()
     }
 }
